--- src/components/review/ProjectChip/style.ts
@@ -1,8 +1,0 @@
-import styled from '@emotion/native';
-
-export const Container = styled.View`
-  width: fit-content;
-  padding: 6px 12px;
-  background: ${({ theme }) => theme.color.CoolNeutral['98']};
-  border-radius: 4px;
-`;
